chore(cleanup): move test request profiles
