+++ Turner.Infrastructure.Crud.Tests/CreateRequestTests.cs
@@ -1,8 +1,11 @@
-﻿using NUnit.Framework;
+﻿using AutoMapper;
+using NUnit.Framework;
 using System.Linq;
 using System.Threading.Tasks;
+using Turner.Infrastructure.Crud.Configuration;
 using Turner.Infrastructure.Crud.Requests;
 using Turner.Infrastructure.Crud.Tests.Fakes;
+using Turner.Infrastructure.Mediator.Decorators;
 
 namespace Turner.Infrastructure.Crud.Tests
 {
@@ -102,6 +105,77 @@ namespace Turner.Infrastructure.Crud.Tests
             Assert.AreEqual(response.Data.Id, Context.Set<User>().First().Id);
             Assert.AreEqual("PreMessage", response.Data.PreMessage);
             Assert.AreEqual("PostMessage/Entity", response.Data.PostMessage);
+        }
+    }
+
+    [DoNotValidate]
+    public class CreateUserWithResponseRequest : UserDto, ICreateRequest<User, UserGetDto>
+    { }
+
+    [DoNotValidate]
+    public class CreateUserWithoutResponseRequest : ICreateRequest<User>
+    {
+        public UserDto User { get; set; }
+    }
+
+    [DoNotValidate]
+    public class DerivedCreateUserWithoutResponseRequest : CreateUserWithoutResponseRequest
+    {
+        public object OtherStuff { get; set; }
+    }
+    
+    public class CreateRequestProfile : CrudRequestProfile<ICreateRequest>
+    {
+        public CreateRequestProfile()
+        {
+            ForEntity<IHasPreMessage>()
+                .BeforeCreating(request =>
+                {
+                    if (request is IHasPreMessage withMessage)
+                        withMessage.PreMessage += "/Entity";
+
+                    return Task.CompletedTask;
+                });
+
+            ForEntity<User>()
+                .BeforeCreating(request =>
+                {
+                    if (request is UserDto dto)
+                        dto.PreMessage += "/User";
+
+                    return Task.CompletedTask;
+                });
+        }
+    }
+
+    public class CreateUserWithoutResponseProfile : CrudRequestProfile<CreateUserWithoutResponseRequest>
+    {
+        public CreateUserWithoutResponseProfile()
+        {
+            ForEntity<User>()
+                .CreateWith(request => Mapper.Map<User>(request.User))
+                .AfterCreating(user =>
+                {
+                    user.PostMessage += "/User";
+                    return Task.CompletedTask;
+                });
+        }
+    }
+
+    public class DefaultCreateRequestProfile<TEntity, TIn> 
+        : CrudRequestProfile<CreateRequest<TEntity, TIn>>
+        where TEntity : class
+    {
+        public DefaultCreateRequestProfile()
+        {
+            ForEntity<TEntity>()
+                .AfterCreating(entity =>
+                {
+                    if (entity is IEntity ent)
+                        ent.PostMessage = "Default";
+
+                    return Task.CompletedTask;
+                });
         }
     }
 }
